refactor: move Paths package to internal
