+++ internal/paths/paths.go
