+++ internal/trace/trace.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/airbytehq/abctl/internal/build"
-	"github.com/airbytehq/abctl/internal/cmd/local/paths"
+	"github.com/airbytehq/abctl/internal/paths"
 	"github.com/airbytehq/abctl/internal/telemetry"
 	"github.com/getsentry/sentry-go"
 	sentryotel "github.com/getsentry/sentry-go/otel"
